requirements.txt
# torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
accelerate==1.5.2
pytorch-lightning==2.5.2
diffusers==0.25.1
transformers==4.31.0
safetensors==0.5.3
xformers==0.0.24
peft==0.10.0
omegaconf==2.3.0
opencv-python==4.11.0.86
pillow==11.0.0
scikit-image==0.25.2
albumentations==2.0.8
kornia==0.8.1
open-clip-torch==2.20.0
openai-clip==1.0.1
timm==1.0.15
lpips==0.1.4
pyiqa==0.1.13
einops==0.8.1
facexlib==0.3.0
numpy==1.26.0
pandas==1.5.3
scipy==1.15.2
datasets==3.4.1
huggingface-hub==0.24.6